+++ src/phenotypic/_core/_pipeline_parts/_image_pipeline_core.py
@@ -35,7 +35,7 @@ from phenotypic.analysis.abc_._set_analyzer import SetAnalyzer
 # Import the entry type from the qc._recipe *submodule* (not the package
 # __init__) so the edge stays ``_core -> qc._recipe -> analysis.abc_`` and
 # never pulls in qc._runner / _cli / gui at module load.
-from phenotypic.qc._recipe import QcRecipeEntry
+from phenotypic.tools_._qc_recipe import QcRecipeEntry
 from phenotypic.tools_.mixin import LazyWidgetMixin
 
 logger = logging.getLogger("ImagePipeline")
